services/inference-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "inference-server"
version = "0.1.0"
description = "gRPC TimesFM / Zeitreihen-Foundation-Model Inference (entkoppelt)."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.112",
  "grpcio>=1.71,<2",
  "httpx>=0.27",
  "numpy>=1.26,<3",
  "protobuf>=5.26,<6",
  "pydantic>=2.6",
  "pydantic-settings>=2",
  "python-json-logger>=2.0.7",
  "sentence-transformers>=3.0.0",
  "torch>=2.2.0",
  "uvicorn>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["src/inference_server"]

[tool.ruff]
line-length = 88
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
